Create an md5 hash of file, and sign it
