+++ app/hooks/useCreateEntity/useCreateTrack.ts
@@ -1,13 +1,14 @@
 /* External dependencies */
 import { useState, ChangeEvent, useEffect } from 'react';
 import { transactions, cryptography } from '@liskhq/lisk-client';
+import md5 from 'md5';
 
 /* Internal dependencies */
 import { useAccount } from '~/hooks/useAccount/useAccount';
 import { MODULES, COMMANDS, FEEDBACK_MESSAGES } from './constants';
-import { AUDIO_CREATE_SCHEMA } from '~/constants/schemas';
 import { CHAIN_ID } from '~/constants/app';
 import { Method, DryRunTxResponse, PostTxResponse } from '~/context/socketContext/types';
+import { AUDIO_CREATE_SCHEMA } from './schemas';
 import { useWS } from '../useWS/useWS';
 import { ValidationStatus } from './types';
 import { validate } from './validator';
@@ -51,8 +52,22 @@ export const useCreateTrack = () => {
   };
 
   const broadcast = async () => {
+    if (!files) {
+      return false;
+    }
     // update account state
     const data = await updateAccount();
+
+    // Get file hash
+    const fileContent = await files[0].arrayBuffer();
+    const md5Hash = md5(new Uint8Array(fileContent)); // Takes around 0.001 ms
+    const { message: hash } = cryptography.ed.signMessageWithPrivateKey(
+      md5Hash, Buffer.from(data.privateKey, 'hex'),
+    ); // Takes around 350 ms
+
+    // @todo request storage info from Streamer
+    const meta = hash;
+
     // Create blockchain transaction and broadcast it
     const tx = {
       module: MODULES.AUDIO,
@@ -63,6 +78,8 @@ export const useCreateTrack = () => {
         name,
         releaseYear,
         artistName,
+        hash,
+        meta,
         genre: [genre],
         collectionID: Buffer.from(collectionID, 'hex'),
         owners: [{
